+++ public/warning.js
@@ -1,2 +1,0 @@
-console.log("%cNÃO COLE NADA AQUI, SENÃO VOCÊ PODE SER HACKEADO", "font-size: 30px;");
-console.log("%cScorpio não irá responsabiliza por sua conta ser hackeada, apenas se for roubada por Brute force", "font-size: 20px;");